requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.16.0
azure-search-documents==11.4.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cryptography==42.0.7
cymem==2.0.8
dataclasses-json==0.6.6
distlib==0.3.7
distro==1.9.0
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
Faker==25.2.0
filelock==3.13.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isodate==0.6.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-experimental==0.0.58
langchain-openai==0.1.7
langchain-text-splitters==0.0.1
langcodes==3.4.0
langsmith==0.1.57
language_data==1.2.0
marisa-trie==1.1.1
MarkupSafe==2.1.5
marshmallow==3.21.2
msal==1.28.0
msal-extensions==1.1.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
mysqlclient==2.2.4
numpy==1.26.4
openai==1.30.1
orjson==3.10.3
packaging==23.2
phonenumbers==8.13.37
pipenv==2023.10.24
platformdirs==3.11.0
portalocker==2.8.2
preshed==3.0.9
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==306
PyYAML==6.0.1
regex==2024.5.10
requests==2.31.0
requests-file==2.1.0
setuptools==68.2.2
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
sqlparse==0.5.0
srsly==2.4.8
tenacity==8.3.0
thinc==8.2.3
tiktoken==0.7.0
tldextract==5.1.2
tqdm==4.66.4
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.24.6
wasabi==1.1.2
weasel==0.3.4
yarl==1.9.4
